# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/retailer1/order/{id}": {
            "get": {
                "description": "读取订单详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "零售商1-Retailer1"
                ],
                "summary": "读取订单详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dao.OrderInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/retailer1/product/{id}": {
            "get": {
                "description": "读取货物批次详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "零售商1-Retailer1"
                ],
                "summary": "读取货物批次详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "批次编号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dao.ProductInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/retailer1/sign/{id}": {
            "post": {
                "description": "签收货物",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "零售商1-Retailer1"
                ],
                "summary": "签收货物",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateLogisticRecordInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "物流追踪ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/retailer1/tracking/{id}": {
            "get": {
                "description": "读取物流详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "零售商1-Retailer1"
                ],
                "summary": "读取物流详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "物流号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dao.LogisticsRecord"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/retailer2/order/{id}": {
            "get": {
                "description": "读取订单详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "零售商2-Retailer2"
                ],
                "summary": "读取订单详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dao.OrderInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/retailer2/product/{id}": {
            "get": {
                "description": "读取货物批次详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "零售商2-Retailer2"
                ],
                "summary": "读取货物批次详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "批次编号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dao.ProductInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/retailer2/sign/{id}": {
            "post": {
                "description": "签收货物",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "零售商2-Retailer2"
                ],
                "summary": "签收货物",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateLogisticRecordInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "物流追踪ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/retailer2/tracking/{id}": {
            "get": {
                "description": "读取物流详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "零售商2-Retailer2"
                ],
                "summary": "读取物流详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "物流号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dao.LogisticsRecord"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dao.LogisticsRecord": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dao.RecordSubItem"
                    }
                }
            }
        },
        "dao.OrderInfo": {
            "type": "object",
            "properties": {
                "batch_number": {
                    "description": "OrderNumber    string  `json:\"order_number\"`",
                    "type": "string"
                },
                "client": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "sorter": {
                    "description": "分拣员",
                    "type": "string"
                },
                "tracking_number": {
                    "type": "string"
                },
                "unit_price": {
                    "type": "number"
                }
            }
        },
...
        "dao.ProductInfo": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "last_modified": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                }
            }
        },
        "dao.RecordSubItem": {
            "type": "object",
            "properties": {
                "record_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateLogisticRecordInput": {
            "type": "object",
            "properties": {
                "operator": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tracking_num": {
                    "type": "string"
                }
            }
        },
        "middleware.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "errmsg": {
                    "type": "string"
                },
                "errno": {
                    "type": "integer"
                },
                "stack": {
                    "type": "object"
                },
                "trace_id": {
                    "type": "object"
                }
            }
        }
    }
}